+++ R/truncated-normal-distribution.R
@@ -9,7 +9,7 @@
 #' @param p	              vector of probabilities.
 #' @param n	              number of observations. If \code{length(n) > 1},
 #'                        the length is taken to be the number required.
-#' @param mu,sigma        location and scale parameters. Scale must be positive.
+#' @param mean,sd         location and scale parameters. Scale must be positive.
 #' @param a,b             minimal and maximal boundries for truncation
 #'                        (\code{-Inf} and \code{Inf} by default).
 #' @param log,log.p	      logical; if TRUE, probabilities p are given as log(p).
@@ -62,6 +62,35 @@
 #' hist(ptnorm(x, 5, 3, b = 7))
 #' plot(ecdf(x))
 #' lines(xx, ptnorm(xx, 5, 3, b = 7), col = "red", lwd = 2)
+#' 
+#' R <- 1e5
+#' partmp <- par(mfrow = c(2,4), mar = c(2,2,2,2))
+#' 
+#' hist(rtnorm(R), freq= FALSE, main = "", xlab = "", ylab = "")
+#' lines(xx, dtnorm(xx), col = "red")
+#' 
+#' hist(rtnorm(R, a = 0), freq= FALSE, main = "", xlab = "", ylab = "")
+#' lines(xx, dtnorm(xx, a = 0), col = "red")
+#' 
+#' hist(rtnorm(R, b = 0), freq= FALSE, main = "", xlab = "", ylab = "")
+#' lines(xx, dtnorm(xx, b = 0), col = "red")
+#' 
+#' hist(rtnorm(R, a = 0, b = 1), freq= FALSE, main = "", xlab = "", ylab = "")
+#' lines(xx, dtnorm(xx, a = 0, b = 1), col = "red")
+#' 
+#' hist(rtnorm(R, a = -1, b = 0), freq= FALSE, main = "", xlab = "", ylab = "")
+#' lines(xx, dtnorm(xx, a = -1, b = 0), col = "red")
+#' 
+#' hist(rtnorm(R, mean = -6, a = 0), freq= FALSE, main = "", xlab = "", ylab = "")
+#' lines(xx, dtnorm(xx, mean = -6, a = 0), col = "red")
+#' 
+#' hist(rtnorm(R, mean = 8, b = 0), freq= FALSE, main = "", xlab = "", ylab = "")
+#' lines(xx, dtnorm(xx, mean = 8, b = 0), col = "red")
+#' 
+#' hist(rtnorm(R, a = 3, b = 5), freq= FALSE, main = "", xlab = "", ylab = "")
+#' lines(xx, dtnorm(xx, a = 3, b = 5), col = "red")
+#' 
+#' par(partmp)
 #'
 #' @name TruncNormal
 #' @aliases TruncNormal
@@ -70,32 +99,32 @@
 #'
 #' @export
 
-dtnorm <- function(x, mu = 0, sigma = 1, a = -Inf, b = Inf, log = FALSE) {
-  .Call('extraDistr_cpp_dtnorm', PACKAGE = 'extraDistr', x, mu, sigma, a, b, log)
+dtnorm <- function(x, mean = 0, sd = 1, a = -Inf, b = Inf, log = FALSE) {
+  cpp_dtnorm(x, mean, sd, a, b, log)
 }
 
 
 #' @rdname TruncNormal
 #' @export
 
-ptnorm <- function(q, mu = 0, sigma = 1, a = -Inf, b = Inf, lower.tail = TRUE, log.p = FALSE) {
-  .Call('extraDistr_cpp_ptnorm', PACKAGE = 'extraDistr', q, mu, sigma, a, b, lower.tail, log.p)
+ptnorm <- function(q, mean = 0, sd = 1, a = -Inf, b = Inf, lower.tail = TRUE, log.p = FALSE) {
+  cpp_ptnorm(q, mean, sd, a, b, lower.tail, log.p)
 }
 
 
 #' @rdname TruncNormal
 #' @export
 
-qtnorm <- function(p, mu = 0, sigma = 1, a = -Inf, b = Inf, lower.tail = TRUE, log.p = FALSE) {
-  .Call('extraDistr_cpp_qtnorm', PACKAGE = 'extraDistr', p, mu, sigma, a, b, lower.tail, log.p)
+qtnorm <- function(p, mean = 0, sd = 1, a = -Inf, b = Inf, lower.tail = TRUE, log.p = FALSE) {
+  cpp_qtnorm(p, mean, sd, a, b, lower.tail, log.p)
 }
 
 
 #' @rdname TruncNormal
 #' @export
 
-rtnorm <- function(n, mu = 0, sigma = 1, a = -Inf, b = Inf) {
+rtnorm <- function(n, mean = 0, sd = 1, a = -Inf, b = Inf) {
   if (length(n) > 1) n <- length(n)
-  .Call('extraDistr_cpp_rtnorm', PACKAGE = 'extraDistr', n, mu, sigma, a, b)
+  cpp_rtnorm(n, mean, sd, a, b)
 }