+++ src/mixture-of-normal-distributions.cpp
@@ -38,23 +38,40 @@ NumericVector cpp_dmixnorm(
   if (k != mu.ncol() || k != sigma.ncol())
     Rcpp::stop("sizes of 'mu', 'sigma', and 'alpha' do not match");
   
-  bool wrong_param;
+  bool wrong_param, missings;
   double alpha_tot;
   
   for (int i = 0; i < Nmax; i++) {
     wrong_param = false;
     alpha_tot = 0.0;
     p[i] = 0.0;
+    missings = false;
+    
     for (int j = 0; j < k; j++) {
-      p[i] += alpha(i % na, j) * R::dnorm(x[i], mu(i % nm, j), sigma(i % ns, j), false);
-      alpha_tot += alpha(i % na, j);
-      if (sigma(i % ns, j) <= 0.0)
+      if (ISNAN(alpha(i % na, j)) || ISNAN(mu(i % nm, j)) || ISNAN(sigma(i % ns, j))) {
+        missings = true;
+        break;
+      }
+      if (alpha(i % na, j) < 0.0 || sigma(i % ns, j) <= 0.0) {
         wrong_param = true;
+        break;
+      }
+      alpha_tot += alpha(i % na, j);
     }
-    if (!tol_equal(alpha_tot, 1.0) || wrong_param) {
+    
+    if (missings || ISNAN(x[i])) {
+      p[i] = NA_REAL;
+      continue;
+    }
+    
+    if (wrong_param) {
       Rcpp::warning("NaNs produced");
       p[i] = NAN;
+      continue;
     }
+    
+    for (int j = 0; j < k; j++)
+      p[i] += (alpha(i % na, j) / alpha_tot) * R::dnorm(x[i], mu(i % nm, j), sigma(i % ns, j), false);
   }
   
   if (log_prob)
@@ -85,23 +102,40 @@ NumericVector cpp_pmixnorm(
   if (k != mu.ncol() || k != sigma.ncol())
     Rcpp::stop("sizes of 'mu', 'sigma', and 'alpha' do not match");
   
-  bool wrong_param;
+  bool wrong_param, missings;
   double alpha_tot;
   
   for (int i = 0; i < Nmax; i++) {
     wrong_param = false;
     alpha_tot = 0.0;
     p[i] = 0.0;
+    missings = false;
+    
     for (int j = 0; j < k; j++) {
-      p[i] += alpha(i % na, j) * R::pnorm(x[i], mu(i % nm, j), sigma(i % ns, j), lower_tail, false);
-      alpha_tot += alpha(i % na, j);
-      if (sigma(i % ns, j) <= 0.0)
+      if (ISNAN(alpha(i % na, j)) || ISNAN(mu(i % nm, j)) || ISNAN(sigma(i % ns, j))) {
+        missings = true;
+        break;
+      }
+      if (alpha(i % na, j) < 0.0 || sigma(i % ns, j) < 0.0) {
         wrong_param = true;
+        break;
+      }
+      alpha_tot += alpha(i % na, j);
     }
-    if (!tol_equal(alpha_tot, 1.0) || wrong_param) {
+    
+    if (missings || ISNAN(x[i])) {
+      p[i] = NA_REAL;
+      continue;
+    }
+    
+    if (wrong_param) {
       Rcpp::warning("NaNs produced");
       p[i] = NAN;
+      continue;
     }
+    
+    for (int j = 0; j < k; j++)
+      p[i] += (alpha(i % na, j) / alpha_tot) * R::pnorm(x[i], mu(i % nm, j), sigma(i % ns, j), lower_tail, false);
   }
   
   if (log_prob)
@@ -130,8 +164,8 @@ NumericVector cpp_rmixnorm(
     Rcpp::stop("sizes of 'mu', 'sigma', and 'alpha' do not match");
   
   int jj;
-  bool wrong_param;
-  double u, p_tmp;
+  bool wrong_param, missings;
+  double u, p_tmp, alpha_tot;
   NumericVector prob(k);
   
   for (int i = 0; i < n; i++) {
@@ -139,35 +173,41 @@ NumericVector cpp_rmixnorm(
     wrong_param = false;
     u = rng_unif();
     p_tmp = 1.0;
+    alpha_tot = 0.0;
+    missings = false;
     
-    for (int j = k-1; j >= 0; j--) {
-      p_tmp -= alpha(i % na, j);
-      if (sigma(i % ns, j) < 0.0 || alpha(i % na, j) < 0.0 || alpha(i % na, j) > 1.0) {
+    for (int j = 0; j < k; j++) {
+      if (ISNAN(alpha(i % na, j)) || ISNAN(mu(i % nm, j)) || ISNAN(sigma(i % ns, j))) {
+        missings = true;
+        break;
+      }
+      if (alpha(i % na, j) < 0.0 || sigma(i % ns, j) < 0.0) {
         wrong_param = true;
         break;
       }
+      alpha_tot += alpha(i % na, j);
+    }
+    
+    if (missings || ISNAN(x[i])) {
+      x[i] = NA_REAL;
+      continue;
+    }
+    
+    if (wrong_param) {
+      Rcpp::warning("NaNs produced");
+      x[i] = NAN;
+      continue;
+    }
+    
+    for (int j = k-1; j >= 0; j--) {
+      p_tmp -= alpha(i % na, j) / alpha_tot;
       if (u > p_tmp) {
         jj = j;
         break;
       }
     }
-    
-    if (!wrong_param && jj > 0) {
-      for (int j = jj-1; j >= 0; j--) {
-        p_tmp -= alpha(i % na, j);
-        if (sigma(i % ns, j) < 0.0 || alpha(i % na, j) < 0.0 || alpha(i % na, j) > 1.0) {
-          wrong_param = true;
-          break;
-        }
-      } 
-    }
 
-    if (wrong_param || !tol_equal(p_tmp, 0.0)) {
-      Rcpp::warning("NaNs produced");
-      x[i] = NAN;
-    } else {
-      x[i] = R::rnorm(mu(i % nm, jj), sigma(i % ns, jj)); 
-    }
+    x[i] = R::rnorm(mu(i % nm, jj), sigma(i % ns, jj)); 
   }
   
   return x;