+++ src/mixture-of-poisson-distributions.cpp
@@ -36,23 +36,40 @@ NumericVector cpp_dmixpois(
   if (k != lambda.ncol())
     Rcpp::stop("sizes of 'lambda' and 'alpha' do not match");
   
-  bool wrong_param;
+  bool wrong_param, missings;
   double alpha_tot;
   
   for (int i = 0; i < Nmax; i++) {
     wrong_param = false;
     alpha_tot = 0.0;
     p[i] = 0.0;
+    missings = false;
+    
     for (int j = 0; j < k; j++) {
-      p[i] += alpha(i % na, j) * R::dpois(x[i], lambda(i % nl, j), false);
-      alpha_tot += alpha(i % na, j);
-      if (lambda(i % nl, j) < 0.0)
+      if (ISNAN(alpha(i % na, j)) || ISNAN(lambda(i % nl, j))) {
+        missings = true;
+        break;
+      }
+      if (alpha(i % na, j) < 0.0 || lambda(i % nl, j) < 0.0) {
         wrong_param = true;
+        break;
+      }
+      alpha_tot += alpha(i % na, j);
     }
-    if (!tol_equal(alpha_tot, 1.0) || wrong_param) {
+    
+    if (missings || ISNAN(x[i])) {
+      p[i] = NA_REAL;
+      continue;
+    }
+    
+    if (wrong_param) {
       Rcpp::warning("NaNs produced");
       p[i] = NAN;
+      continue;
     }
+    
+    for (int j = 0; j < k; j++)
+      p[i] += (alpha(i % na, j) / alpha_tot) * R::dpois(x[i], lambda(i % nl, j), false);
   }
   
   if (log_prob)
@@ -81,23 +98,40 @@ NumericVector cpp_pmixpois(
   if (k != lambda.ncol())
     Rcpp::stop("sizes of 'lambda' and 'alpha' do not match");
   
-  bool wrong_param;
+  bool wrong_param, missings;
   double alpha_tot;
   
   for (int i = 0; i < Nmax; i++) {
     wrong_param = false;
     alpha_tot = 0.0;
     p[i] = 0.0;
+    missings = false;
+    
     for (int j = 0; j < k; j++) {
-      p[i] += alpha(i % na, j) * R::ppois(x[i], lambda(i % nl, j), lower_tail, false);
-      alpha_tot += alpha(i % na, j);
-      if (lambda(i % nl, j) < 0.0)
+      if (ISNAN(alpha(i % na, j)) || ISNAN(lambda(i % nl, j))) {
+        missings = true;
+        break;
+      }
+      if (alpha(i % na, j) < 0.0 || lambda(i % nl, j) < 0.0) {
         wrong_param = true;
+        break;
+      }
+      alpha_tot += alpha(i % na, j);
     }
-    if (!tol_equal(alpha_tot, 1.0) || wrong_param) {
+    
+    if (missings || ISNAN(x[i])) {
+      p[i] = NA_REAL;
+      continue;
+    }
+    
+    if (wrong_param) {
       Rcpp::warning("NaNs produced");
       p[i] = NAN;
+      continue;
     }
+    
+    for (int j = 0; j < k; j++)
+      p[i] += (alpha(i % na, j) / alpha_tot) * R::ppois(x[i], lambda(i % nl, j), lower_tail, false);
   }
   
   if (log_prob)
@@ -124,8 +158,8 @@ NumericVector cpp_rmixpois(
     Rcpp::stop("sizes of 'lambda' and 'alpha' do not match");
   
   int jj;
-  bool wrong_param;
-  double u, p_tmp;
+  bool wrong_param, missings;
+  double u, p_tmp, alpha_tot;
   NumericVector prob(k);
   
   for (int i = 0; i < n; i++) {
@@ -133,33 +167,41 @@ NumericVector cpp_rmixpois(
     wrong_param = false;
     u = rng_unif();
     p_tmp = 1.0;
+    alpha_tot = 0.0;
+    missings = false;
     
-    for (int j = k-1; j >= 0; j--) {
-      p_tmp -= alpha(i % na, j);
-      if (lambda(i % nl, j) < 0.0 || alpha(i % na, j) < 0.0 || alpha(i % na, j) > 1.0) {
+    for (int j = 0; j < k; j++) {
+      if (ISNAN(alpha(i % na, j)) || ISNAN(lambda(i % nl, j))) {
+        missings = true;
+        break;
+      }
+      if (alpha(i % na, j) < 0.0 || lambda(i % nl, j) < 0.0) {
         wrong_param = true;
         break;
       }
+      alpha_tot += alpha(i % na, j);
+    }
+    
+    if (missings || ISNAN(x[i])) {
+      x[i] = NA_REAL;
+      continue;
+    }
+    
+    if (wrong_param) {
+      Rcpp::warning("NaNs produced");
+      x[i] = NAN;
+      continue;
+    }
+    
+    for (int j = k-1; j >= 0; j--) {
+      p_tmp -= alpha(i % na, j) / alpha_tot;
       if (u > p_tmp) {
         jj = j;
         break;
       }
     }
     
-    if (!wrong_param && jj > 0) {
-      for (int j = jj-1; j >= 0; j--) {
-        p_tmp -= alpha(i % na, j);
-        if (lambda(i % nl, j) < 0.0 || alpha(i % na, j) < 0.0 || alpha(i % na, j) > 1.0)
-          wrong_param = true;
-      } 
-    }
-    
-    if (wrong_param || !tol_equal(p_tmp, 0.0)) {
-      Rcpp::warning("NaNs produced");
-      x[i] = NAN;
-    } else {
-      x[i] = R::rpois(lambda(i % nl, jj)); 
-    }
+    x[i] = R::rpois(lambda(i % nl, jj)); 
   }
   
   return x;