+++ src/shared.cpp
@@ -18,7 +18,7 @@ bool isInteger(double x) {
   return true;
 }
 
-// Dealing with Inf
+// Dealing with Inf and NAs
 
 bool anyFinite(Rcpp::NumericVector x) {
   int n = x.length();
@@ -36,6 +36,14 @@ double finite_max(Rcpp::NumericVector x) {
       max_x = x[i];
   }
   return max_x;
+}
+
+bool allNA(Rcpp::NumericVector x) {
+  int n = x.length();
+  for (int i = 0; i < n; i++)
+    if (!ISNAN(x[i]))
+      return false;
+  return true;
 }
 
 // Standard normal
@@ -74,6 +82,8 @@ double rng_unif() {
 }
 
 double rng_bern(double p) {
+  if (ISNAN(p))
+    return NA_REAL;
   if (p < 0.0 || p > 1.0) {
     Rcpp::warning("NaNs produced");
     return NAN;
@@ -86,5 +96,39 @@ double rng_sign() {
   double u = rng_unif();
   return (u > 0.5) ? 1.0 : -1.0;
 }
- 
- 
+
+// Checking parameters
+
+Rcpp::NumericMatrix normalize_prob(const Rcpp::NumericMatrix& prob) {
+  
+  int n = prob.nrow();
+  int k = prob.ncol();
+  double p_tot;
+  bool wrong_param;
+  Rcpp::NumericMatrix p = Rcpp::clone(prob);
+  
+  for (int i = 0; i < n; i++) {
+    wrong_param = false;
+    p_tot = 0.0;
+    
+    for (int j = 0; j < k; j++) {
+      if (p(i, j) < 0.0 || ISNAN(p(i, j))) {
+        wrong_param = true;
+        break;
+      }
+      p_tot += p(i, j);
+    }
+    
+    if (wrong_param) {
+      Rcpp::warning("NaNs produced");
+      for (int j = 0; j < k; j++)
+        p(i, j) = NAN;
+    } else if (p_tot > 1.0) {
+      for (int j = 0; j < k; j++)
+        p(i, j) /= p_tot;
+    }
+  }
+  
+  return p;
+}
+