+++ src/shared.h
@@ -7,10 +7,11 @@
 bool tol_equal(double x, double y);
 bool isInteger(double x);
 
-// Dealing with Inf
+// Dealing with Inf and NAs
 
 bool anyFinite(Rcpp::NumericVector x);
 double finite_max(Rcpp::NumericVector x);
+bool allNA(Rcpp::NumericVector x);
 
 // Standard normal
 
@@ -28,5 +29,9 @@ double lfactorial(double x);
 double rng_unif();         // standard uniform
 double rng_bern(double p); // Bernoulli distribution
 double rng_sign();         // Rademacher distribution
+
+// Checking parameters
+
+Rcpp::NumericMatrix normalize_prob(const Rcpp::NumericMatrix& prob);
 
 #endif